all: Provide nozzle ID and zone to MetricHandler
This change introduces a configuration option to provide an ID to
uniquely identify the instance running the nozzle. Although this may
be configured via the env, configuration will be overriden by ID
and zone if the nozzle is running on GCE.

The ID and zone are used as labels in metrics sent by the heartbeater.