stackdriver: Fix problems with AutoCulledMetricsBuffer
This change fixes two problems when a larger volume of metrics (~100/s)
are processed:

1. `metric_sink` would range over a map of metrics and pass the pointer
of the local map value to a method. Because the pointer is re-used
within the loop, the downstream value could change. This manifests as
duplicate metrics appearing at high throughput. A pointer to the map
value is now passed.

2. A hash is calculated to uniquely identify each metric. This hash was
generated by ranging over a map and appending strings. Because order
isn't consistent, it was possible for duplicate metrics to exist.