heartbeat: Set correct event time
This change sets the event time to the current time each time a metric
batch is sent. Previously, the time defaulted to a time in the past,
which allowed the first batch to be inserted, but subsequent batches to
be rejected because their event time was earlier than existing point.