Break requirement for credentials job in tile
Move the responsibility of handling application_default_credentials to
the stackdriver-nozzle job. The job now populates its own creds file
from the same job spec (credentials.application_..) and pipes it into
the nozzle via the ENV.

Removed the step in the credentials job that used to do this for the
nozzle.

Fixes #88