nozzle: handle multiple calls to Stop
The done channel was being left open after the consumer drained the done
meesage. Calling the Stop() method 2 times in this state would cause the
program to hang while trying to send done <- struct{}{}.

Change the nozzle to keep track of running state and to return an error
when it's stopped incorrectly.

See #112 for more detail