metrics_router: Route messages after the buffer
Utilize the AutoCullingMetricsBuffer by placing the router after it in
the pipeline. This allows the buffer to do it's thing and sample metrics
to reduce the amount we log to Stackdriver Logging.

If this was in the nozzle or written as part of a sink we'd need to
perform the buffering/culling at that point or maintain two seperate
methods for doing it.