metrics_buffer: Don't block new messages during flush
Background:
The mutex on mb.metrics blocks calls to PostMetric(s). This prevents the
nozzle from draining messages from loggregator which drops the
connection.

This was an issue when we only wrote to Stackdriver Monitoring but now
that we also write to Stackdriver Logging this issue has been magnified.

Fix:
- Only keep the mutex in the critical section where we copy and flush the
  internal buffer.

Testing:
- Without this change the nozzle consistently folds in ~4mins on my
  machine and and utilizes 2% CPU.
- With this change it appears more stable and the usage goes up to 5%
  which means less blocking!