stackdriver/metric: measure out of order write errors
Once there is more than a single nozzle writing we expect these errors
and should not count them in the same way. These occur when more than a
single nozzle is writing to stackdriver. If nozzle A writes a metric at
time T2 and nozzle B writes a netric at time T1 then nozzle B will get
an error that it is writing out of order.

- Emit `metrics.errors.out_of_order` and `metrics.errors.unknown` to
  help operators understand this.
- Don't output the entire metric request for these errors. That detail
  is useful for tracking down specific errors and this error is tracked
  down.

Fixes #98