move metrics batch to stackdriver.MetricAdapter
The batch sizing isn't general enough to be apart of the
metric_pipeline. Each endpoint we write to (Logging, Monitoring) has its
own batching requirements. The Logging SDK already handles batching for
us but the Monitoring SDK does not.

Batching in the AutoCulledMetricsBuffer made sense when it delt with
individual Metrics instead of MetricEvents. It does not make sense now
that a MetricEvent is made of many Metrics as it does not have the
knowledge on how to batch requests to the endpoints.

There's a few follow ups before merging this as a PR.

fixes #137