Prepend event origin to all generated metric names.
[Pivotal's docs][1] use the origin to differentiate between metrics that
have the same name, explicitly prepending it to the metric name. The
DataDog nozzle does the same thing. Consistency with these is valuable,
though this does introduce a change in metric naming that is not
backwards-compatible.

Example name changes, first the "now matches the docs, but whatever":

*   AuctioneerTaskAuctionsFailed -> auctioneer.AuctioneerTaskAuctionsFailed
*   LRPsExtra -> bbs.LRPsExtra

Then the "wow, disambiguation is helpful!":

*   RequestLatency -> bbs.RequestLatency
*   latency -> gorouter.latency
*   Latency -> etcd.Latency

*   requests.completed -> cc.requests.completed
*   total_requests -> gorouter.total_requests
*   ReceivingRequestRate -> etcd.ReceivingRequestRate

Then the "maybe we don't want to do this, ugh":

*   numGoRoutines ->
    *   tps_listener.numGoRoutines
    *   tcp_emitter.numGoRoutines
    *   ... etc

The last one is different because numGoRoutines means the same thing
across multiple origins. Using a label to differentiate instead here
is probably the right thing to do, but conditionally appending the
origin requires more complexity to enumerate where that condition holds.

[1]: https://docs.pivotal.io/pivotalcf/1-12/monitoring/kpi.html