heartbeater: new heartbeater implementation
This implementation guarentees that Increment is non-blocking. This is
achieved by having the emit call grab the mutex to the data before it
grabs a handle on it, then resets the data and returns the mutex. No
mater how slow emit() takes it will not block Increment() for longer
than it takes to create a new map.

This design is also considerably simpler. No log handler is needed and
the running goroutine is no longer responsible for flushing, stopping,
and keeping track of incoming Increment calls.